# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todolist App API",
    "version": "1.0.0",
    "description": "Actual REST API contract for the current backend implementation."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Invalid request." },
          "details": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          }
        },
        "required": ["code", "message"]
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        },
        "required": ["message"]
      },
...
      "Team": {
        "type": "object",
        "properties": {
          "teamId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "createdBy": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["teamId", "name", "createdBy", "createdAt", "updatedAt"]
      },
      "TeamRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 100 }
        },
        "required": ["name"]
      },
      "TeamMember": {
        "type": "object",
        "properties": {
          "teamMemberId": { "type": "string", "format": "uuid" },
          "teamId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["ADMIN", "MEMBER", "VIEWER"] },
          "joinedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["teamMemberId", "teamId", "userId", "role", "joinedAt"]
      },
      "TeamInvitation": {
        "type": "object",
        "properties": {
          "invitationId": { "type": "string", "format": "uuid" },
          "teamId": { "type": "string", "format": "uuid" },
          "invitedUserId": { "type": "string", "format": "uuid" },
          "invitedBy": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["MEMBER", "VIEWER"] },
          "status": { "type": "string", "enum": ["PENDING", "ACCEPTED", "DECLINED", "EXPIRED"] },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "respondedAt": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["invitationId", "teamId", "invitedUserId", "invitedBy", "role", "status", "expiresAt", "createdAt"]
      },
      "TeamInvitationCreateRequest": {
        "type": "object",
        "properties": {
          "invitedUserId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["MEMBER", "VIEWER"] }
        },
        "required": ["invitedUserId", "role"]
      },
      "TeamRoleUpdateRequest": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["ADMIN", "MEMBER", "VIEWER"] }
        },
        "required": ["role"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/teams": {
      "get": {
        "tags": ["Teams"],
        "summary": "List teams for current user",
        "responses": {
          "200": { "description": "Team list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Team" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Teams"],
        "summary": "Create team",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeamRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Team" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}": {
      "get": {
        "tags": ["Teams"],
        "summary": "Get team",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Team", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Team" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Teams"],
        "summary": "Update team",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeamRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated team", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Team" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Teams"],
        "summary": "Delete team",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}/members": {
      "get": {
        "tags": ["Teams"],
        "summary": "List team members",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Member list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TeamMember" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}/members/{userId}/role": {
      "patch": {
        "tags": ["Teams"],
        "summary": "Change member role",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeamRoleUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated member", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamMember" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "Business rule violation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}/members/me": {
      "delete": {
        "tags": ["Teams"],
        "summary": "Leave team",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Left team" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "Business rule violation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}/members/{userId}": {
      "delete": {
        "tags": ["Teams"],
        "summary": "Kick member",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Member removed" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/teams/{teamId}/invitations": {
      "get": {
        "tags": ["Teams"],
        "summary": "List team invitations",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Invitation list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TeamInvitation" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Teams"],
        "summary": "Create team invitation",
        "parameters": [
          { "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeamInvitationCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Invitation created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamInvitation" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Invitation or membership conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/invitations/{invitationId}/accept": {
      "patch": {
        "tags": ["Teams"],
        "summary": "Accept invitation",
        "parameters": [
          { "name": "invitationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Already a member", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "Invitation not acceptable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/invitations/{invitationId}/decline": {
      "patch": {
        "tags": ["Teams"],
        "summary": "Decline invitation",
        "parameters": [
          { "name": "invitationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Declined", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "Invitation not declineable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}